Updated workflows and node e2e tests to node20
