+++ .github/workflows/run-e2e-tests-dotnet3-lcon.yaml
@@ -6,71 +6,71 @@ on:
       - dev
       - releases/*
     paths-ignore:
-      - '**.md'
+      - "**.md"
   pull_request:
     paths-ignore:
-      - '**.md'
+      - "**.md"
 
 env:
   AZURE_FUNCTIONAPP_NAME: gae-fa-dotnet3-lcon
-  AZURE_FUNCTIONAPP_PACKAGE_PATH: './tests/e2e/dotnet3'
-  DOTNET_VERSION: '3.1.403'
+  AZURE_FUNCTIONAPP_PACKAGE_PATH: "./tests/e2e/dotnet3"
+  DOTNET_VERSION: "3.1.403"
 
 jobs:
   run:
     name: Run E2E Tests
     runs-on: ubuntu-latest
     steps:
-    - name: Checkout
-      uses: actions/checkout@v2
+      - name: Checkout
+        uses: actions/checkout@v4
 
-    - name: Set Node.js 16.x for GitHub Action
-      uses: actions/setup-node@v1
-      with:
-        node-version: 16.x
+      - name: Set Node.js 20.x for GitHub Action
+        uses: actions/setup-node@v4
+        with:
+          node-version: 20.x
 
-    - name: Run Npm Install for GitHub Action
-      run: npm install
+      - name: Run Npm Install for GitHub Action
+        run: npm install
 
-    - name: Build GitHub Action
-      run: npm run build
+      - name: Build GitHub Action
+        run: npm run build
 
-    - name: E2E Setup DotNet ${{ env.DOTNET_VERSION }} Environment
-      uses: actions/setup-dotnet@v1
-      with:
-        dotnet-version: ${{ env.DOTNET_VERSION }}
+      - name: E2E Setup DotNet ${{ env.DOTNET_VERSION }} Environment
+        uses: actions/setup-dotnet@v4
+        with:
+          dotnet-version: ${{ env.DOTNET_VERSION }}
 
-    - name: E2E Resolve Project Dependencies Using Dotnet
-      shell: pwsh
-      run: |
-        pushd '${{ env.AZURE_FUNCTIONAPP_PACKAGE_PATH }}'
-        dotnet build --configuration Release --output ./output
-        echo "$env:GITHUB_SHA" > ./output/sha.txt
-        popd
+      - name: E2E Resolve Project Dependencies Using Dotnet
+        shell: pwsh
+        run: |
+          pushd '${{ env.AZURE_FUNCTIONAPP_PACKAGE_PATH }}'
+          dotnet build --configuration Release --output ./output
+          echo "$env:GITHUB_SHA" > ./output/sha.txt
+          popd
 
-    - name: 'Login via Azure CLI'
-      uses: azure/login@v1
-      with:
-        creds: ${{ secrets.RBAC_GAE_FA_DOTNET3_LCON }}
+      - name: "Login via Azure CLI"
+        uses: azure/login@v2
+        with:
+          creds: ${{ secrets.RBAC_GAE_FA_DOTNET3_LCON }}
 
-    - name: E2E Run Azure Functions Action
-      uses: ./
-      id: fa
-      with:
-        app-name: ${{ env.AZURE_FUNCTIONAPP_NAME }}
-        package: '${{ env.AZURE_FUNCTIONAPP_PACKAGE_PATH }}/output'
+      - name: E2E Run Azure Functions Action
+        uses: ./
+        id: fa
+        with:
+          app-name: ${{ env.AZURE_FUNCTIONAPP_NAME }}
+          package: "${{ env.AZURE_FUNCTIONAPP_PACKAGE_PATH }}/output"
 
-    - name: E2E Check HttpTrigger Result
-      shell: pwsh
-      run: |
-        $i = 0
-        while ($i -lt 10) {
-          sleep 10
-          $RESPONSE = $(curl "${{ steps.fa.outputs.app-url }}/api/HttpTrigger")
-          $RESULT = ($RESPONSE -eq "$env:GITHUB_SHA")
-          if ($RESULT) {
-            exit 0
+      - name: E2E Check HttpTrigger Result
+        shell: pwsh
+        run: |
+          $i = 0
+          while ($i -lt 10) {
+            sleep 10
+            $RESPONSE = $(curl "${{ steps.fa.outputs.app-url }}/api/HttpTrigger")
+            $RESULT = ($RESPONSE -eq "$env:GITHUB_SHA")
+            if ($RESULT) {
+              exit 0
+            }
+            $i = $i + 1
           }
-          $i = $i + 1
-        }
-        exit 1
+          exit 1
