--- .github/workflows/run-e2e-tests-node16-wcon.yaml
@@ -1,79 +1,0 @@
-name: RUN_E2E_TESTS_NODE16_WCON
-on:
-  push:
-    branches:
-      - master
-      - dev
-      - releases/*
-    paths-ignore:
-      - '**.md'
-  pull_request:
-    paths-ignore:
-      - '**.md'
-
-env:
-  AZURE_FUNCTIONAPP_NAME: gae-fa-node16-wcon
-  AZURE_FUNCTIONAPP_PACKAGE_PATH: './tests/e2e/node16'
-  NODE_VERSION: '16.x'
-
-jobs:
-  run:
-    name: Run E2E Tests
-    runs-on: windows-latest
-    steps:
-    - name: Checkout
-      uses: actions/checkout@v2
-
-    - name: Set Node.js 16.x for GitHub Action
-      uses: actions/setup-node@v1
-      with:
-        node-version: 16.x
-
-    - name: Run Npm Install for GitHub Action
-      run: npm install
-
-    - name: Build GitHub Action
-      run: npm run build
-
-    - name: E2E Setup Node ${{ env.NODE_VERSION }} Environment
-      uses: actions/setup-node@v1
-      with:
-        node-version: ${{ env.NODE_VERSION }}
-
-    - name: E2E Resolve Project Dependencies Using Npm
-      shell: bash
-      run: |
-        pushd '${{ env.AZURE_FUNCTIONAPP_PACKAGE_PATH }}'
-        npm install
-        npm run build --if-present
-        npm run test --if-present
-        echo "$GITHUB_SHA" > sha.txt
-        popd
-
-    - name: 'Login via Azure CLI'
-      uses: azure/login@v1
-      with:
-        creds: ${{ secrets.RBAC_GAE_FA_NODE16_WCON }}
-
-    - name: E2E Run Azure Functions Action
-      uses: ./
-      id: fa
-      with:
-        app-name: ${{ env.AZURE_FUNCTIONAPP_NAME }}
-        package: ${{ env.AZURE_FUNCTIONAPP_PACKAGE_PATH }}
-        scm-do-build-during-deployment: true
-
-    - name: E2E Check HttpTrigger Result
-      shell: pwsh
-      run: |
-        $i = 0
-        while ($i -lt 10) {
-          sleep 10
-          $RESPONSE = $(curl "${{ steps.fa.outputs.app-url }}/api/HttpTrigger")
-          $RESULT = ($RESPONSE -eq "$env:GITHUB_SHA")
-          if ($RESULT) {
-            exit 0
-          }
-          $i = $i + 1
-        }
-        exit 1
