+++ .github/workflows/run-e2e-tests-node20-lcon.yaml
@@ -1,4 +1,4 @@
-name: RUN_E2E_TESTS_DOTNET3_WCON
+name: RUN_E2E_TESTS_NODE20_LCON
 on:
   push:
     branches:
@@ -12,14 +12,14 @@ on:
       - "**.md"
 
 env:
-  AZURE_FUNCTIONAPP_NAME: gae-fa-dotnet3-wcon
-  AZURE_FUNCTIONAPP_PACKAGE_PATH: "./tests/e2e/dotnet3"
-  DOTNET_VERSION: "3.1.403"
+  AZURE_FUNCTIONAPP_NAME: gae-fa-node20-lcon
+  AZURE_FUNCTIONAPP_PACKAGE_PATH: "./tests/e2e/node20"
+  NODE_VERSION: "20.x"
 
 jobs:
   run:
     name: Run E2E Tests
-    runs-on: windows-latest
+    runs-on: ubuntu-latest
     steps:
       - name: Checkout
         uses: actions/checkout@v4
@@ -35,30 +35,32 @@ jobs:
       - name: Build GitHub Action
         run: npm run build
 
-      - name: E2E Setup DotNet ${{ env.DOTNET_VERSION }} Environment
-        uses: actions/setup-dotnet@v4
+      - name: E2E Setup Node ${{ env.NODE_VERSION }} Environment
+        uses: actions/setup-node@v4
         with:
-          dotnet-version: ${{ env.DOTNET_VERSION }}
+          node-version: ${{ env.NODE_VERSION }}
 
-      - name: E2E Resolve Project Dependencies Using Dotnet
-        shell: pwsh
+      - name: E2E Resolve Project Dependencies Using Npm
+        shell: bash
         run: |
           pushd '${{ env.AZURE_FUNCTIONAPP_PACKAGE_PATH }}'
-          dotnet build --configuration Release --output ./output
-          echo "$env:GITHUB_SHA" > ./output/sha.txt
+          npm install
+          npm run build --if-present
+          npm run test --if-present
+          echo "$GITHUB_SHA" > sha.txt
           popd
 
       - name: "Login via Azure CLI"
         uses: azure/login@v2
         with:
-          creds: ${{ secrets.RBAC_GAE_FA_DOTNET3_WCON }}
+          creds: ${{ secrets.RBAC_GAE_FA_NODE16_LCON }}
 
       - name: E2E Run Azure Functions Action
         uses: ./
         id: fa
         with:
           app-name: ${{ env.AZURE_FUNCTIONAPP_NAME }}
-          package: "${{ env.AZURE_FUNCTIONAPP_PACKAGE_PATH }}/output"
+          package: ${{ env.AZURE_FUNCTIONAPP_PACKAGE_PATH }}
 
       - name: E2E Check HttpTrigger Result
         shell: pwsh
