+++ .github/workflows/run-e2e-tests-powershell6-wcon.yaml
@@ -6,64 +6,64 @@ on:
       - dev
       - releases/*
     paths-ignore:
-      - '**.md'
+      - "**.md"
   pull_request:
     paths-ignore:
-      - '**.md'
+      - "**.md"
 
 env:
   AZURE_FUNCTIONAPP_NAME: gae-fa-powershell6-wcon
-  AZURE_FUNCTIONAPP_PACKAGE_PATH: './tests/e2e/powershell6'
+  AZURE_FUNCTIONAPP_PACKAGE_PATH: "./tests/e2e/powershell6"
 
 jobs:
   run:
     name: Run E2E Tests
     runs-on: windows-latest
     steps:
-    - name: Checkout
-      uses: actions/checkout@v2
+      - name: Checkout
+        uses: actions/checkout@v4
 
-    - name: Set Node.js 16.x for GitHub Action
-      uses: actions/setup-node@v1
-      with:
-        node-version: 16.x
+      - name: Set Node.js 20.x for GitHub Action
+        uses: actions/setup-node@v4
+        with:
+          node-version: 20.x
 
-    - name: Run Npm Install for GitHub Action
-      run: npm install
+      - name: Run Npm Install for GitHub Action
+        run: npm install
 
-    - name: Build GitHub Action
-      run: npm run build
+      - name: Build GitHub Action
+        run: npm run build
 
-    - name: E2E Generate Unique Token in sha.txt
-      shell: bash
-      run: |
-        pushd '${{ env.AZURE_FUNCTIONAPP_PACKAGE_PATH }}'
-        echo "$GITHUB_SHA" > sha.txt
-        popd
+      - name: E2E Generate Unique Token in sha.txt
+        shell: bash
+        run: |
+          pushd '${{ env.AZURE_FUNCTIONAPP_PACKAGE_PATH }}'
+          echo "$GITHUB_SHA" > sha.txt
+          popd
 
-    - name: 'Login via Azure CLI'
-      uses: azure/login@v1
-      with:
-        creds: ${{ secrets.RBAC_GAE_FA_POWERSHELL6_WCON }}
+      - name: "Login via Azure CLI"
+        uses: azure/login@v2
+        with:
+          creds: ${{ secrets.RBAC_GAE_FA_POWERSHELL6_WCON }}
 
-    - name: E2E Run Azure Functions Action
-      uses: ./
-      id: fa
-      with:
-        app-name: ${{ env.AZURE_FUNCTIONAPP_NAME }}
-        package: ${{ env.AZURE_FUNCTIONAPP_PACKAGE_PATH }}
+      - name: E2E Run Azure Functions Action
+        uses: ./
+        id: fa
+        with:
+          app-name: ${{ env.AZURE_FUNCTIONAPP_NAME }}
+          package: ${{ env.AZURE_FUNCTIONAPP_PACKAGE_PATH }}
 
-    - name: E2E Check HttpTrigger Result
-      shell: pwsh
-      run: |
-        $i = 0
-        while ($i -lt 10) {
-          sleep 10
-          $RESPONSE = $(curl "${{ steps.fa.outputs.app-url }}/api/HttpTrigger")
-          $RESULT = ($RESPONSE -eq "$env:GITHUB_SHA")
-          if ($RESULT) {
-            exit 0
+      - name: E2E Check HttpTrigger Result
+        shell: pwsh
+        run: |
+          $i = 0
+          while ($i -lt 10) {
+            sleep 10
+            $RESPONSE = $(curl "${{ steps.fa.outputs.app-url }}/api/HttpTrigger")
+            $RESULT = ($RESPONSE -eq "$env:GITHUB_SHA")
+            if ($RESULT) {
+              exit 0
+            }
+            $i = $i + 1
           }
-          $i = $i + 1
-        }
-        exit 1
+          exit 1
