+++ .github/workflows/run-e2e-tests-python310-lcon-msi.yaml
@@ -6,77 +6,77 @@ on:
       - dev
       - releases/*
     paths-ignore:
-      - '**.md'
+      - "**.md"
   pull_request:
     paths-ignore:
-      - '**.md'
+      - "**.md"
 
 permissions:
   id-token: write
-  
+
 env:
   AZURE_FUNCTIONAPP_NAME: gae-fa-python310-lcon-msi
-  AZURE_FUNCTIONAPP_PACKAGE_PATH: './tests/e2e/python310'
-  PYTHON_VERSION: '3.10'
+  AZURE_FUNCTIONAPP_PACKAGE_PATH: "./tests/e2e/python310"
+  PYTHON_VERSION: "3.10"
 
 jobs:
   run:
     name: Run E2E Tests
     runs-on: ubuntu-latest
     steps:
-    - name: Checkout
-      uses: actions/checkout@v2
+      - name: Checkout
+        uses: actions/checkout@v4
 
-    - name: Set Node.js 16.x for GitHub Action
-      uses: actions/setup-node@v1
-      with:
-        node-version: 16.x
+      - name: Set Node.js 20.x for GitHub Action
+        uses: actions/setup-node@v4
+        with:
+          node-version: 20.x
 
-    - name: Setup Python ${{ env.PYTHON_VERSION }} Environment
-      uses: actions/setup-python@v1
-      with:
-        python-version: ${{ env.PYTHON_VERSION }}
+      - name: Setup Python ${{ env.PYTHON_VERSION }} Environment
+        uses: actions/setup-python@v1
+        with:
+          python-version: ${{ env.PYTHON_VERSION }}
 
-    - name: Run Npm Install for GitHub Action
-      run: npm install
+      - name: Run Npm Install for GitHub Action
+        run: npm install
 
-    - name: Build GitHub Action
-      run: npm run build
+      - name: Build GitHub Action
+        run: npm run build
 
-    - name: E2E Resolve Project Dependencies Using Pip
-      shell: bash
-      run: |
-        pushd '${{ env.AZURE_FUNCTIONAPP_PACKAGE_PATH }}'
-        python -m pip install --upgrade pip
-        pip install -r requirements.txt --target=".python_packages/lib/site-packages"
-        echo "$GITHUB_SHA" > sha.txt
-        popd
-    
-    - name: 'Login via Azure CLI'
-      uses: azure/login@v1
-      with:
-        client-id: ${{ secrets.AZURE_CLIENT_ID_FA_PYTHON310_LCON_MSI }}
-        tenant-id: ${{ secrets.AZURE_TENANT_ID_FA_PYTHON310_LCON_MSI }}
-        subscription-id: ${{ secrets.AZURE_SUBSCRIPTION_ID_FA_PYTHON310_LCON_MSI }}
+      - name: E2E Resolve Project Dependencies Using Pip
+        shell: bash
+        run: |
+          pushd '${{ env.AZURE_FUNCTIONAPP_PACKAGE_PATH }}'
+          python -m pip install --upgrade pip
+          pip install -r requirements.txt --target=".python_packages/lib/site-packages"
+          echo "$GITHUB_SHA" > sha.txt
+          popd
 
-    - name: E2E Run Azure Functions Action
-      uses: ./
-      id: fa
-      with:
-        app-name: ${{ env.AZURE_FUNCTIONAPP_NAME }}
-        package: ${{ env.AZURE_FUNCTIONAPP_PACKAGE_PATH }}
+      - name: "Login via Azure CLI"
+        uses: azure/login@v2
+        with:
+          client-id: ${{ secrets.AZURE_CLIENT_ID_FA_PYTHON310_LCON_MSI }}
+          tenant-id: ${{ secrets.AZURE_TENANT_ID_FA_PYTHON310_LCON_MSI }}
+          subscription-id: ${{ secrets.AZURE_SUBSCRIPTION_ID_FA_PYTHON310_LCON_MSI }}
 
-    - name: E2E Check HttpTrigger Result
-      shell: pwsh
-      run: |
-        $i = 0
-        while ($i -lt 10) {
-          sleep 10
-          $RESPONSE = $(curl "${{ steps.fa.outputs.app-url }}/api/HttpTrigger")
-          $RESULT = ($RESPONSE -eq "$env:GITHUB_SHA")
-          if ($RESULT) {
-            exit 0
+      - name: E2E Run Azure Functions Action
+        uses: ./
+        id: fa
+        with:
+          app-name: ${{ env.AZURE_FUNCTIONAPP_NAME }}
+          package: ${{ env.AZURE_FUNCTIONAPP_PACKAGE_PATH }}
+
+      - name: E2E Check HttpTrigger Result
+        shell: pwsh
+        run: |
+          $i = 0
+          while ($i -lt 10) {
+            sleep 10
+            $RESPONSE = $(curl "${{ steps.fa.outputs.app-url }}/api/HttpTrigger")
+            $RESULT = ($RESPONSE -eq "$env:GITHUB_SHA")
+            if ($RESULT) {
+              exit 0
+            }
+            $i = $i + 1
           }
-          $i = $i + 1
-        }
-        exit 1
+          exit 1
