--- tests/e2e/node16/HttpTrigger/function.json
@@ -1,19 +1,0 @@
-{
-  "bindings": [
-    {
-      "authLevel": "Anonymous",
-      "type": "httpTrigger",
-      "direction": "in",
-      "name": "req",
-      "methods": [
-        "get",
-        "post"
-      ]
-    },
-    {
-      "type": "http",
-      "direction": "out",
-      "name": "res"
-    }
-  ]
-}
