--- tests/e2e/node16/sha.txt
@@ -1,1 +1,0 @@
-sha-txt-node16-placeholder
