+++ tests/e2e/node20/HttpTrigger/index.js
@@ -1,0 +1,9 @@
+var lodash = require('lodash');
+var fs = require('fs');
+
+module.exports = async function (context, req) {
+    const folder = context.executionContext.functionDirectory;
+    const filePath = fs.realpathSync(`${folder}/../sha.txt`);
+    const text = fs.readFileSync(filePath,'utf8');
+    context.res = { body: text };
+}
